+++ drivers/hubspot/internal/base_stream.go
