+++ drivers/hubspot/internal/config.go
@@ -1,4 +1,4 @@
-package models
+package driver
 
 import (
 	"fmt"
