+++ drivers/postgres/internal/models.go
@@ -1,4 +1,4 @@
-package models
+package driver
 
 import (
 	"fmt"
@@ -105,4 +105,15 @@ func (c *Config) ToConnectionString() string {
 	}
 
 	return connStr
+}
+
+type Table struct {
+	Schema string `db:"table_schema"`
+	Name   string `db:"table_name"`
+}
+
+type ColumnDetails struct {
+	Name       string  `db:"column_name"`
+	DataType   *string `db:"data_type"`
+	IsNullable *string `db:"is_nullable"`
 }
