+++ drivers/s3/internal/models.go
@@ -1,4 +1,4 @@
-package models
+package driver
 
 import (
 	"fmt"
