+++ drivers/s3/internal/reader/parquet.go
