+++ drivers/s3/internal/reader/reader.go
